style(api): rename `root.ts` for clearer filename semantics
