+++ packages/api/test/sync-state.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { appRouter } from "../src/root";
+import { appRouter } from "../src/app-router";
 import { getContext } from "./helpers";
 
 describe("Sync state route", async () => {
